Replace Rake loading and invocation in rails_helper
Make `db:setup` a precondition of the default `rake` task.

This *is* different than `db:reset` in one way I specifically intend.
It does not drop the entire database. Reasons dropping the DB is
unnecessary:
- in CI, the DB will already have just been freshly created
- testing uses its own DB, so anything from dev will not affect it
- all examples are run inside a database transaction, so it should be
difficult and uncommon for tests to bleed DB data across each other.

The main thing is just to make sure the seed data is in place for the
examples that are dependent on it, which `db:setup` does.